+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 45
+#define BUILD 46